pyproject.toml
[tool.poetry]
name = "bot-starter"
version = "0.1.0"
description = "A starter bot repo to be forked or cloned for future projects"
authors = ["kaj <40004347+KAJdev@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
naff = "^1.2.0"
aiohttp = "^3.8.1"
motor = "^3.0.0"
pymongo = "^4.1.1"
srv = "^1.1.0"
python-dotenv = "^0.20.0"
dacite = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
